requirements.txt
pygame==2.6.1
neat-python==0.92